oldrequirements.txt
#fastapi
#uvicorn[standard]
#python-dotenv
#vonage>=4.1.0

#vonage-voice==1.4.0



##################



fastapi
uvicorn[standard]
python-dotenv
vonage>=4.1.0
vonage-messages
#==1.2.0
vonage-voice==1.4.0
google-genai
Jinja2

google-generativeai
python-multipart

#vonage-utils

